\name{gadget_likelihood_component}
\alias{gadget_likelihood_component}

\title{Gadget likelihood components}
\description{
    Structures representing a component of a GADGET likelihood file.
}

\usage{
gadget_likelihood_component(type, weight = 0, name = type, likelihoodfile = 'likelihood', ...)
}

\arguments{
    \item{type}{
        Type of group to create. One of penalty, understocking, catchstatistics,
        catchdistribution, stockdistribution.
    }
    \item{name}{
        A descriptive name for the component
    }
    \item{weight}{
        A numeric weighting
    }
    \item{likelihoodfile}{
        The likelihood file this component should end up in
    }
    \item{...}{
        Extra parameters for the group. See details.
    }
}

\details{
In addition, \code{penalty} understands:
\describe{
    \item{data}{A \code{data.frame} with 2 columns, "switch" and "power"}
}

\code{catchstatistics} understands:
\describe{
    \item{data_function}{The function Gadget should use, by default guesses based on the function that generated \code{data}}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{list} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{list} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
}

\code{catchdistribution} understands:
\describe{
    \item{data_function}{The function Gadget should use, by default uses sumofsquares}
    \item{data_function_params}{Extra parameters to supply to gadget, based on the function}
    \item{aggregationlevel}{TRUE or FALSE, defaults to FALSE}
    \item{overconsumption}{TRUE or FALSE, defaults to FALSE}
    \item{epsilon}{Numeric, defaults to 10}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{list} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{list} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{length}{An \code{list} of lengths, taken from \code{attr(data, "length")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
}

\code{stockdistribution} understands:
\describe{
    \item{data_function}{The function Gadget should use, by default uses sumofsquares}
    \item{overconsumption}{TRUE or FALSE, defaults to FALSE}
    \item{epsilon}{Numeric, defaults to 10}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{list} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{list} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{length}{An \code{list} of lengths, taken from \code{attr(data, "length")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
}

\code{surveyindices} understands:
\describe{
    \item{sitype}{What data the component is based on, see GADGET manual}
    \item{biomass}{0 or 1, defaults to 0}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{list} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{list} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{length}{An \code{list} of lengths, taken from \code{attr(data, "length")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
    \item{surveynames}{List of acoustic survey names}
    \item{fittype, slope, intercept}{Fit options, see GADGET manual}
}

\code{stomachcontent} understands:
\describe{
    \item{data_function}{Function GADGET will use}
    \item{epsilon}{To be used when calculated probability is low}
    \item{prey_labels}{Names for each of the }
    \item{prey_digestion_coefficients}{Optional. Digestion coefficients to put in prey aggregation file}
    \item{predator_names}{Vector of predator stock names}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
}

\code{migrationpenalty} understands:
\describe{
    \item{stockname}{Stock name}
    \item{powercoeffs}{2 power coefficients}
}

}

\examples{
# Any example could be added to a file with the following:-
\dontrun{
gd <- gadget_directory('./out')
gadget_dir_write(gd, component)
}

# Create a penalty component
gadget_likelihood_component("penalty",
        name = "bounds",
        weight = "0.5",
        data = data.frame(
            switch = c("default"),
            power = c(2),
            stringsAsFactors = FALSE))

# Create an understocking component
gadget_likelihood_component("understocking", name ="understock")
}
