\name{mfdb-package}
\alias{mfdb-package}
\docType{package}

\title{MareFrame DB querying library}
\description{
  Tools to query a MareFrame DB and reformat results in forms
  useful for GADGET and EwE models.
}
\details{
  For a list of documentation use \code{library(help = "mfdb")}
}

\author{
  Maintainer: Jamie Lentin <jamie.lentin@shuttlethread.com>
}
