\name{gadget_likelihood_component}
\alias{gadget_likelihood_component}

\title{Gadget likelihood components}
\description{
    Structures representing a component of a GADGET likelihood file.
}

\usage{
gadget_likelihood_component(type, ...)
}

\arguments{
    \item{type}{
        Type of group to create. One of penalty, understocking, catchstatistics,
        catchdistribution, stockdistribution.
    }
    \item{...}{
        Extra parameters for the group. See details.
    }
}

\details{
The arguments you can supply will depend on the the component you are creating.
All will understand the following
\describe{
    \item{name}{A descriptive name for the component}
    \item{weight}{A numeric weighting}
}

In addition, \code{penalty} understands:
\describe{
    \item{data}{A \code{data.frame} with 2 columns, "switch" and "power"}
}

\code{catchstatistics} understands:
\describe{
    \item{data_function}{The function Gadget should use, by default guesses based on the function that generated \code{data}}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{mfdb_group} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{mfdb_group} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
}

\code{catchdistribution} understands:
\describe{
    \item{data_function}{The function Gadget should use, by default uses sumofsquares}
    \item{data_function_params}{Extra parameters to supply to gadget, based on the function}
    \item{aggregationlevel}{TRUE or FALSE, defaults to FALSE}
    \item{overconsumption}{TRUE or FALSE, defaults to FALSE}
    \item{epsilon}{Numeric, defaults to 10}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{mfdb_group} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{mfdb_group} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{length}{An \code{mfdb_group} of lengths, taken from \code{attr(data, "length")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
}

\code{stockdistribution} understands:
\describe{
    \item{data_function}{The function Gadget should use, by default uses sumofsquares}
    \item{overconsumption}{TRUE or FALSE, defaults to FALSE}
    \item{epsilon}{Numeric, defaults to 10}
    \item{data}{A \code{data.frame} probably generated by \code{mfdb_sample_meanlength_stddev}}
    \item{area}{An \code{mfdb_group} of areas, taken from \code{attr(data, "area")} if not supplied}
    \item{age}{An \code{mfdb_group} of ages, taken from \code{attr(data, "age")} if not supplied}
    \item{length}{An \code{mfdb_group} of lengths, taken from \code{attr(data, "length")} if not supplied}
    \item{fleetnames}{List of fleet names}
    \item{stocknames}{List of stock names}
}

}

\examples{
# Any example could be added to a file with the following:-
\dontrun{
gd <- gadget_directory('./out')
gadget_dir_write(gd, component)
}

# Create a penalty component
gadget_likelihood_component("penalty",
        name = "bounds",
        weight = "0.5",
        data = data.frame(
            switch = c("default"),
            power = c(2),
            stringsAsFactors = FALSE))

# Create an understocking component
gadget_likelihood_component("understocking", name ="understock")
}
